[refactor/#179] signup 관련 구조 변경
