--- resource-server/src/main/java/com/inhabas/api/domain/questionaire/usecase/QuestionnaireService.java
@@ -1,10 +1,0 @@
-package com.inhabas.api.domain.questionaire.usecase;
-
-import com.inhabas.api.domain.questionaire.dto.QuestionnaireDto;
-
-import java.util.List;
-
-public interface QuestionnaireService {
-
-    List<QuestionnaireDto> getQuestionnaire();
-}
