+++ resource-server/src/main/java/com/inhabas/api/domain/questionnaire/usecase/QuestionnaireService.java
@@ -1,0 +1,10 @@
+package com.inhabas.api.domain.questionnaire.usecase;
+
+import com.inhabas.api.domain.questionnaire.dto.QuestionnaireDto;
+
+import java.util.List;
+
+public interface QuestionnaireService {
+
+    List<QuestionnaireDto> getQuestionnaire();
+}
