+++ resource-server/src/main/java/com/inhabas/api/domain/signUp/domain/entity/Answer.java
@@ -1,8 +1,8 @@
-package com.inhabas.api.domain.member.domain.entity;
+package com.inhabas.api.domain.signUp.domain.entity;
 
 import com.inhabas.api.auth.domain.oauth2.member.domain.entity.Member;
 import com.inhabas.api.domain.BaseEntity;
-import com.inhabas.api.domain.questionaire.domain.Questionnaire;
+import com.inhabas.api.domain.questionnaire.domain.Questionnaire;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
