+++ resource-server/src/main/java/com/inhabas/api/domain/signUp/repository/AnswerRepository.java
@@ -1,6 +1,6 @@
-package com.inhabas.api.domain.member.repository;
+package com.inhabas.api.domain.signUp.repository;
 
-import com.inhabas.api.domain.member.domain.entity.Answer;
+import com.inhabas.api.domain.signUp.domain.entity.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
@@ -9,5 +9,4 @@ public interface AnswerRepository extends JpaRepository<Answer, Long> {
 
     List<Answer> findByMember_Id(Long memberId);
 
-    boolean existsByMember_Id(Long memberId);
 }
