+++ resource-server/src/main/java/com/inhabas/api/domain/signUpSchedule/domain/usecase/SignUpAvailabilityCheckerImpl.java
@@ -1,4 +1,4 @@
-package com.inhabas.api.domain.signUpSchedule.domain;
+package com.inhabas.api.domain.signUpSchedule.domain.usecase;
 
 import com.inhabas.api.domain.signUpSchedule.dto.SignUpScheduleDto;
 import java.time.LocalDateTime;
