+++ resource-server/src/main/java/com/inhabas/api/domain/signUpSchedule/domain/usecase/SignUpSchedulerStrict.java
@@ -1,4 +1,4 @@
-package com.inhabas.api.domain.signUpSchedule.domain;
+package com.inhabas.api.domain.signUpSchedule.domain.usecase;
 
 import com.inhabas.api.domain.signUpSchedule.domain.entity.SignUpSchedule;
 import com.inhabas.api.domain.signUpSchedule.dto.SignUpScheduleDto;
