+++ resource-server/src/main/java/com/inhabas/api/web/SignUpController.java
@@ -1,10 +1,10 @@
 package com.inhabas.api.web;
 
 import com.inhabas.api.auth.domain.oauth2.majorInfo.dto.MajorInfoDto;
-import com.inhabas.api.domain.member.dto.AnswerDto;
-import com.inhabas.api.domain.member.dto.SignUpDto;
-import com.inhabas.api.domain.member.usecase.SignUpService;
-import com.inhabas.api.domain.questionaire.dto.QuestionnaireDto;
+import com.inhabas.api.domain.signUp.dto.AnswerDto;
+import com.inhabas.api.domain.signUp.dto.SignUpDto;
+import com.inhabas.api.domain.signUp.usecase.SignUpService;
+import com.inhabas.api.domain.questionnaire.dto.QuestionnaireDto;
 import com.inhabas.api.web.argumentResolver.Authenticated;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
