+++ resource-server/src/main/java/com/inhabas/api/web/interceptor/SignUpControllerInterceptor.java
@@ -1,6 +1,6 @@
 package com.inhabas.api.web.interceptor;
 
-import com.inhabas.api.domain.signUpSchedule.domain.SignUpAvailabilityChecker;
+import com.inhabas.api.domain.signUpSchedule.domain.usecase.SignUpAvailabilityChecker;
 import com.inhabas.api.domain.signUpSchedule.SignUpNotAvailableException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.servlet.HandlerInterceptor;
